+++ clone/openrocket-release-24.12.RC.01/core/src/main/java/edu/mit/rocket_team/zephyrus/util/data/RTFudgedAirbrakesData.java
@@ -1,27 +1,27 @@
 package edu.mit.rocket_team.zephyrus.util.data;
 
 public class RTFudgedAirbrakesData extends RTFudgedData {
-    public double altitude;
-    public double vel_z;
-    public double accel_z;
+    public float altitude;
+    public float vel_z;
+    public float accel_z;
     public boolean apogeeReached;
 
 
     public RTFudgedAirbrakesData(double altitude, double vel_z, double accel_z, boolean apogeeReached) {
         super();
-        this.altitude = altitude;
-        this.vel_z = vel_z;
-        this.accel_z = accel_z;
+        this.altitude = (float) altitude;
+        this.vel_z = (float) vel_z;
+        this.accel_z = (float) accel_z;
         this.apogeeReached = apogeeReached;
     }
 
-    public double getAltitude() {
+    public float getAltitude() {
         return altitude;
     }
-    public double getVel_z() {
+    public float getVel_z() {
         return vel_z;
     }
-    public double getAccel_z() {
+    public float getAccel_z() {
         return accel_z;
     }
     public boolean isApogeeReached() {
@@ -32,13 +32,13 @@ public class RTFudgedAirbrakesData extends RTFudgedData {
         this.apogeeReached = apogeeReached;
     }
     public void setAltitude(double altitude) {
-        this.altitude = altitude;
+        this.altitude = (float) altitude;
     }
     public void setVel_z(double vel_z) {
-        this.vel_z = vel_z;
+        this.vel_z = (float) vel_z;
     }
     public void setAccel_z(double accel_z) {
-        this.accel_z = accel_z;
+        this.accel_z = (float) accel_z;
     }
